Update on "[JIT][Symbolic Shape Analysis] expose op shape functions"
fixes https://github.com/pytorch/pytorch/issues/68204

cc silvasean 

[ghstack-poisoned]
